refactor: `locales` name correction
