+++ src/locales/fa.ts
