+++ src/locales/index.ts
